+++ zipline-kotlin-plugin/src/main/kotlin/app/cash/zipline/kotlin/ZiplineIrGenerationExtension.kt
@@ -24,6 +24,8 @@ import org.jetbrains.kotlin.ir.declarations.IrClass
 import org.jetbrains.kotlin.ir.declarations.IrModuleFragment
 import org.jetbrains.kotlin.ir.expressions.IrCall
 import org.jetbrains.kotlin.ir.expressions.IrExpression
+import org.jetbrains.kotlin.ir.types.getClass
+import org.jetbrains.kotlin.ir.util.classId
 import org.jetbrains.kotlin.ir.util.fqNameWhenAvailable
 import org.jetbrains.kotlin.ir.util.isInterface
 
@@ -39,7 +41,7 @@ class ZiplineIrGenerationExtension(
 
         try {
           if (declaration.isInterface &&
-            declaration.superTypes.any { it.classId == ziplineApis.ziplineScopedClassId }
+            declaration.superTypes.any { it.getClass()?.classId == ziplineApis.ziplineScopedClassId }
           ) {
             throw ZiplineCompilationException(
               element = declaration,
@@ -49,7 +51,7 @@ class ZiplineIrGenerationExtension(
           }
 
           if (declaration.isInterface &&
-            declaration.superTypes.any { it.classId == ziplineApis.ziplineServiceClassId }
+            declaration.superTypes.any { it.getClass()?.classId == ziplineApis.ziplineServiceClassId }
           ) {
             AdapterGenerator(
               pluginContext,
