--- zipline-kotlin-plugin/src/main/kotlin/app/cash/zipline/kotlin/utils.kt
@@ -1,35 +1,0 @@
-package app.cash.zipline.kotlin
-
-import org.jetbrains.kotlin.ir.types.IrType
-import org.jetbrains.kotlin.ir.types.classFqName
-import org.jetbrains.kotlin.ir.types.getClass
-import org.jetbrains.kotlin.ir.util.isLocal
-import org.jetbrains.kotlin.ir.util.packageFqName
-import org.jetbrains.kotlin.name.CallableId
-import org.jetbrains.kotlin.name.ClassId
-import org.jetbrains.kotlin.name.FqName
-import org.jetbrains.kotlin.name.Name
-
-val IrType.classId: ClassId?
-  get() {
-    val irClass = getClass()
-    val packageName = irClass?.packageFqName
-    val className = classFqName
-    return if (packageName != null && className != null) {
-      ClassId(packageName, className, irClass.isLocal)
-    } else {
-      null
-    }
-  }
-
-fun FqName.classId(name: String): ClassId {
-  return ClassId(this, Name.identifier(name))
-}
-
-fun FqName.callableId(name: String): CallableId {
-  return CallableId(this, Name.identifier(name))
-}
-
-fun ClassId.callableId(name: String): CallableId {
-  return CallableId(this, Name.identifier(name))
-}
